+++ Backend/src/Modules/Nodes/Nodes.Application/Data/INodesDbContext.cs
@@ -1,7 +1,4 @@
-using Microsoft.EntityFrameworkCore;
-using Nodes.Domain.Models;
-
-namespace Nodes.Application;
+namespace Nodes.Application.Data;
 
 public interface INodesDbContext
 {
