+++ Backend/src/Modules/Nodes/Nodes.Infrastructure/Data/NodesDbContext.cs
@@ -1,4 +1,5 @@
 using System.Reflection;
+using Nodes.Application.Data;
 
 namespace Nodes.Infrastructure.Data;
